new todo task done + delete

+++ src/main/java/annanas_manager/services/impl/TaskServiceImpl.java
@@ -1,14 +1,20 @@
 package annanas_manager.services.impl;
 
 
+import annanas_manager.DTO.CustomUserDTO;
 import annanas_manager.DTO.TaskDTO;
 import annanas_manager.entities.CustomUser;
 import annanas_manager.entities.Task;
+import annanas_manager.entities.enums.TaskStatus;
+import annanas_manager.repositories.CustomUserRepository;
 import annanas_manager.repositories.TaskRepository;
 import annanas_manager.services.TaskService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
+import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
 @Service
@@ -16,25 +22,51 @@ public class TaskServiceImpl implements TaskService{
 
     @Autowired
     private TaskRepository taskRepository;
+    @Autowired
+    private CustomUserRepository userRepository;
 
     @Override
-    public Task add(TaskDTO taskDTO) {
+    public void add(TaskDTO taskDTO) {
+        taskDTO.setCreateDate(new Date(System.currentTimeMillis()));
+        taskDTO.setStatus(TaskStatus.NEW);
         Task task = Task.fromDTO(taskDTO);
-        return taskRepository.saveAndFlush(task);
+        task.setCreatedBy(userRepository.findByEmail(taskDTO.getCreatedBy().getEmail()));
+        System.out.println(task);
+        taskRepository.saveAndFlush(task);
     }
 
     @Override
-    public void delete(long id) {
-        taskRepository.delete(id);
+    public boolean delete(long id, String email) {
+        Task task = taskRepository.findById(id);
+        CustomUser user = userRepository.findByEmail(email);
+        if (task.getCreatedBy().equals(user)){
+            taskRepository.delete(id);
+            return true;
+        }
+        return false;
     }
 
     @Override
-    public Task edit(Task task) {
-        return taskRepository.saveAndFlush(task);
+    public void edit(TaskDTO taskDTO) {
+        Task task = Task.fromDTO(taskDTO);
+        taskRepository.saveAndFlush(task);
     }
 
     @Override
-    public List<Task> getAll(CustomUser id) {
-        return taskRepository.findByUsername(id);
+    @Transactional(readOnly=true)
+    public CustomUserDTO findUser(String email) {
+        CustomUser user = userRepository.findByEmail(email);
+        return user.toDTO();
+    }
+
+    @Override
+    public List<TaskDTO> findByUser(CustomUserDTO customUserDTO) {
+        CustomUser customUser = userRepository.findByEmail(customUserDTO.getEmail());
+        List<Task> tasks = taskRepository.findByUsername(customUser);
+        List<TaskDTO> taskDTOs = new ArrayList<>();
+        for (Task task:tasks) {
+            taskDTOs.add(task.toDTO());
+        }
+        return taskDTOs;
     }
 }
